Revert "[asm] move irq to user registers code"
This reverts commit bca2429a8f35153a1681e7987745e7282629bf5a.

+++ include/asm/irq.h
@@ -7,7 +7,7 @@
 
 #include <kernel/types.h>
 
-ALIAS(irqflags_t, uint32_t);
+ALIAS(irqflags_t, unsigned long);
 
 void irq_local_init(void);
 void irq_local_enable(void);
